refactor: moves example to examples/
